==> notebooks/04_hyperparameter_tuning.ipynb <==
## Hyperparameter Tuning

import joblib
import datetime
import pandas as pd
from sklearn.model_selection import GridSearchCV, RandomizedSearchCV
from sklearn.pipeline import Pipeline
from sklearn.feature_selection import SelectKBest, f_classif, SequentialFeatureSelector
from sklearn.neural_network import MLPClassifier
from sklearn.svm import SVC
from sklearn.ensemble import RandomForestClassifier
from sklearn.neighbors import KNeighborsClassifier
from itertools import product
import time

## Load the training data

# Load split datasets
X_train = joblib.load("../data/splits/X_train.pkl")
X_test = joblib.load("../data/splits/X_test.pkl")
y_train = joblib.load("../data/splits/y_train.pkl")
y_test = joblib.load("../data/splits/y_test.pkl")

print("Train/test split loaded successfully!")

## Models

models = {
    "RF": RandomForestClassifier(random_state=42),
    "SVM": SVC(probability=True),
    "MLP2": MLPClassifier(hidden_layer_sizes=(2,), max_iter=5000, random_state=42),
    "MLP4": MLPClassifier(hidden_layer_sizes=(4,), max_iter=5000, random_state=42),
    "MLP6": MLPClassifier(hidden_layer_sizes=(6,), max_iter=5000, random_state=42),
    "MLP8": MLPClassifier(hidden_layer_sizes=(8,), max_iter=5000, random_state=42),
    "3NN": KNeighborsClassifier(n_neighbors=3),
    "5NN": KNeighborsClassifier(n_neighbors=5),
    "10NN": KNeighborsClassifier(n_neighbors=10),
    "15NN": KNeighborsClassifier(n_neighbors=15),
}

## Search Parameter Grids

# ---- FEATURE SPACE DIMENSION ----

n_features = X_train.shape[1]
k_values = list(range(2, n_features, 2))
print (k_values)

# ---- SEARCH PARAMETER GRIDS ----

param_grids_anova = {
    "RF": {
        "feature_selection__k": k_values,  # k
        "classifier__n_estimators": [50, 100, 200],  # Number of trees
        "classifier__max_depth": [5, 10, 20],  # Tree depth
        "classifier__min_samples_split": [2, 5, 10],  # Minimum samples per split
        "classifier__min_samples_leaf": [1, 2, 4],  # Minimum leaf nodes
        "classifier__max_features": ["sqrt", "log2"]  # Feature subset size at each split
    },
    "MLP2": {
        "feature_selection__k": k_values,  # k
        "classifier__learning_rate_init": [0.001, 0.003, 0.01, 0.03, 0.1],
        "classifier__batch_size": [16, 32, 64, 128],
        "classifier__alpha": [1e-5, 1e-4, 1e-3, 1e-2, 1e-1] # L2 penalty (regularization term) parameter
    },
    "MLP4": {
        "feature_selection__k": k_values,  # k
        "classifier__learning_rate_init": [0.001, 0.003, 0.01, 0.03, 0.1],
        "classifier__batch_size": [16, 32, 64, 128],
        "classifier__alpha": [1e-5, 1e-4, 1e-3, 1e-2, 1e-1] # L2 penalty (regularization term) parameter
    },
    "MLP6": {
        "feature_selection__k": k_values,  # k
        "classifier__learning_rate_init": [0.001, 0.003, 0.01, 0.03, 0.1],
        "classifier__batch_size": [16, 32, 64, 128],
        "classifier__alpha": [1e-5, 1e-4, 1e-3, 1e-2, 1e-1] # L2 penalty (regularization term) parameter
    },
    "MLP8": {
        "feature_selection__k": k_values,
        "classifier__learning_rate_init": [0.001, 0.003, 0.01, 0.03, 0.1],
        "classifier__batch_size": [16, 32, 64, 128],
        "classifier__alpha": [1e-5, 1e-4, 1e-3, 1e-2, 1e-1]
    },
    "SVM": {
        "feature_selection__k": k_values,
        "classifier__C": [0.1, 1, 10, 100],
        "classifier__kernel": ["linear", "rbf", "poly"],
        "classifier__gamma": ["scale", "auto"]
    },
    "3NN": {
        "feature_selection__k": k_values,
        "classifier__weights": ["uniform", "distance"],
        "classifier__metric": ["euclidean", "manhattan", "minkowski"]
    },
    "5NN": {
        "feature_selection__k": k_values,
        "classifier__weights": ["uniform", "distance"],
        "classifier__metric": ["euclidean", "manhattan", "minkowski"]
    },
    "10NN": {
        "feature_selection__k": k_values,
        "classifier__weights": ["uniform", "distance"],
        "classifier__metric": ["euclidean", "manhattan", "minkowski"]
    },
    "15NN": {
        "feature_selection__k": k_values,
        "classifier__weights": ["uniform", "distance"],
        "classifier__metric": ["euclidean", "manhattan", "minkowski"]
    },
}

param_grids_sfs = {
    "RF": {
        "feature_selection__n_features_to_select": k_values,  # n_features_to_select  
        "classifier__n_estimators": [50, 100, 200],  
        "classifier__max_depth": [5, 7, 10, 20],  
        "classifier__min_samples_split": [2, 5, 10],  
        "classifier__min_samples_leaf": [1, 2, 4],
        "classifier__max_features": ["sqrt", "log2", None]
    },
    "MLP2": {
        "feature_selection__n_features_to_select": k_values,
        "classifier__batch_size": [16, 32, 64, 128],
        "classifier__learning_rate_init": [0.001, 0.003, 0.01, 0.03, 0.1],
        "classifier__alpha": [1e-5, 1e-4, 1e-3, 1e-2, 1e-1]
    },
    "MLP4": {
        "feature_selection__n_features_to_select": k_values,
        "classifier__batch_size": [16, 32, 64, 128],
        "classifier__learning_rate_init": [0.001, 0.003, 0.01, 0.03, 0.1],
        "classifier__alpha": [1e-5, 1e-4, 1e-3, 1e-2, 1e-1]
    },
    "MLP6": {
        "feature_selection__n_features_to_select": k_values,
        "classifier__batch_size": [16, 32, 64, 128],
        "classifier__learning_rate_init": [0.001, 0.003, 0.01, 0.03, 0.1],
        "classifier__alpha": [1e-5, 1e-4, 1e-3, 1e-2, 1e-1]
    },
    "MLP8": {
        "feature_selection__n_features_to_select": k_values,
        "classifier__batch_size": [16, 32, 64, 128],
        "classifier__learning_rate_init": [0.001, 0.003, 0.01, 0.03, 0.1],
        "classifier__alpha": [1e-5, 1e-4, 1e-3, 1e-2, 1e-1]
    },
    "SVM": {
        "feature_selection__n_features_to_select": k_values,
        "classifier__C": [0.1, 1, 10, 100],  
        "classifier__kernel": ["linear", "rbf", "poly"],  
        "classifier__gamma": ["scale", "auto"]  
    },
    "3NN": {
        "feature_selection__n_features_to_select": k_values,  
        "classifier__weights": ["uniform", "distance"],  
        "classifier__metric": ["euclidean", "manhattan", "minkowski"]  
    },
    "5NN": {
        "feature_selection__n_features_to_select": k_values,
        "classifier__weights": ["uniform", "distance"],  
        "classifier__metric": ["euclidean", "manhattan", "minkowski"]  
    },
    "10NN": {
        "feature_selection__n_features_to_select": k_values,
        "classifier__weights": ["uniform", "distance"],  
        "classifier__metric": ["euclidean", "manhattan", "minkowski"]  
    },
    "15NN": {
        "feature_selection__n_features_to_select": k_values,
        "classifier__weights": ["uniform", "distance"],  
        "classifier__metric": ["euclidean", "manhattan", "minkowski"]  
    },
}

## Hyperparameter Tuning

best_models = {}
results = []

for name, model in models.items():
    print(f"Tuning hyperparameters for {name}...")

    for fs_method in ["ANOVA", "SFS"]:
        # ---- DEFINE FEATURE SELECTION METHOD ----
        
        if fs_method == "ANOVA":
            selector = SelectKBest(score_func=f_classif)
            param_grid = param_grids_anova[name]
            SearchClass = GridSearchCV
        else:
            selector = SequentialFeatureSelector(model)
            param_grid = param_grids_sfs[name]
            SearchClass = RandomizedSearchCV  # We'll use RandomizedSearch for SFS

        # ---- CREATE PIPELINE ----
        
        pipeline = Pipeline([
            ("feature_selection", selector),
            ("classifier", model)
        ])

        # ---- BUILD SEARCH OBJECT ----
        
        if fs_method == "ANOVA":
            # GridSearchCV
            search = SearchClass(pipeline, param_grid, cv=10, scoring="roc_auc", n_jobs=-1)
        else:
            # RandomizedSearchCV
            total_combinations = len(list(product(*param_grid.values())))
            n_iter_adjusted = min(20, total_combinations)
            search = SearchClass(
                pipeline, param_grid, 
                n_iter=n_iter_adjusted, cv=10, scoring="roc_auc", 
                n_jobs=-1, random_state=42
            )
        
        # ---- MEASURE TIME ----
        
        start_search = time.time()
        search.fit(X_train, y_train)
        end_search = time.time()
        search_time = end_search - start_search

        # ---- REFIT BEST MODEL ----
        
        start_refit = time.time()
        best_model = search.best_estimator_
        end_refit = time.time()
        refit_time = end_refit - start_refit
        best_score = search.best_score_
        best_params = search.best_params_
        
        # ---- SAVE MLP CONVERGENCE ----
        
        mlp_iterations = None
        classifier_step = best_model.named_steps["classifier"]
        if isinstance(classifier_step, MLPClassifier):
            mlp_iterations = classifier_step.n_iter_
        

        # ---- GET SELECTED FEATURES ----
        
        selected_features = X_train.columns[best_model.named_steps["feature_selection"].get_support()]
        selected_features_str = ", ".join(selected_features)

        # ---- STORE RESULTS ----
        
        results.append({
            "Model": name,
            "Feature Selection": fs_method,
            "ROC AUC Score": best_score,
            "Selected Features": selected_features_str,
            "Best Params": best_params,
            "Refit Time (s)": refit_time,
            "Search Time (s)": search_time,
            "Convergence": mlp_iterations
        })

        # ---- STORE BEST MODEL ----
        
        best_models[f"{name} ({fs_method})"] = best_model

        print(f"Best {name} ({fs_method}) ROC_AUC: {best_score:.4f}")
        print(f"Best Params: {best_params}")
        print(f"Selected Features: {selected_features_str}")
        if mlp_iterations is not None:
            print(f"MLP converged in {mlp_iterations} iterations")
        print(f"Final Refit Time: {refit_time:.2f}s")
        print(f"Search Time: {search_time:.2f} s\n")


## Save Data
The best performing models and results are saved to disk for future evaluation.

# ---- SAVE BEST MODELS ----

timestamp = datetime.datetime.now().strftime("%d-%m-%Y-%H-%M")

for model_name, model in best_models.items():
    joblib.dump(model, f"../models/{model_name.replace(' ', '_').replace('(', '').replace(')', '')}_{timestamp}.pkl")

results_df = pd.DataFrame(results)

filename = f"../data/results/04_hyperparameter_tuning_results_{timestamp}.csv"
results_df.to_csv(filename, index=False)

print("Results and best performing models saved for future evaluation.")
